add more tests for dydx python impl
